# openapi/content.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Platform Content API",
        "description": "API сервиса content платформы",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/admin/v1/projects/{project}/content/instructs": {
            "get": {
                "tags": [
                    "content"
                ],
                "summary": "GET /api/admin/v1/projects/{project}/content/instructs",
                "operationId": "content_index_api_admin_v1_projects_project_content_instructs",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "content"
                ],
                "summary": "POST /api/admin/v1/projects/{project}/content/instructs",
                "operationId": "content_store_api_admin_v1_projects_project_content_instructs",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "category",
                                    "rule",
                                    "schema"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 120
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "rule": {
                                        "type": "string",
                                        "maxLength": 20000
                                    },
                                    "schema": {
                                        "type": "object"
                                    },
                                    "published": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/v1/projects/{project}/content/instructs/categories": {
            "get": {
                "tags": [
                    "content"
                ],
                "summary": "GET /api/admin/v1/projects/{project}/content/instructs/categories",
                "operationId": "content_categories_api_admin_v1_projects_project_content_instructs",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/v1/projects/{project}/content/instructs/schema-presets": {
            "get": {
                "tags": [
                    "content"
                ],
                "summary": "GET /api/admin/v1/projects/{project}/content/instructs/schema-presets",
                "operationId": "content_schemaPresets_api_admin_v1_projects_project_content_instructs_schema_presets",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/v1/projects/{project}/content/instructs/{instruct}": {
            "get": {
                "tags": [
                    "content"
                ],
                "summary": "GET /api/admin/v1/projects/{project}/content/instructs/{instruct}",
                "operationId": "content_show_api_admin_v1_projects_project_content_instructs_instruct",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "instruct",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "content"
                ],
                "summary": "PUT /api/admin/v1/projects/{project}/content/instructs/{instruct}",
                "operationId": "content_update_api_admin_v1_projects_project_content_instructs_instruct",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "instruct",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "category",
                                    "rule",
                                    "schema"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "maxLength": 120
                                    },
                                    "category": {
                                        "type": "string"
                                    },
                                    "rule": {
                                        "type": "string",
                                        "maxLength": 20000
                                    },
                                    "schema": {
                                        "type": "object"
                                    },
                                    "published": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "content"
                ],
                "summary": "DELETE /api/admin/v1/projects/{project}/content/instructs/{instruct}",
                "operationId": "content_destroy_api_admin_v1_projects_project_content_instructs_instruct",
                "parameters": [
                    {
                        "name": "project",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "instruct",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            },
            "apiKey": {
                "type": "apiKey",
                "name": "X-Api-Key",
                "in": "header"
            }
        }
    },
...
}